Move install script

+++ bulkhours/core/installer.py
@@ -102,7 +102,7 @@ def main(argv=sys.argv[1:]):
 
     # Install main package
     if is_colab:
-        if args.tokens["atoken"] != DEFAULT_TOKEN:
+        if "atoken" in args.tokens and args.tokens["atoken"] != DEFAULT_TOKEN:
             os.system(
                 f"cd {bulk_dir} && rm -rf bulkhours_admin 2> /dev/null && git clone https://{args.tokens['atoken']}@github.com/guydegnol/bulkhours_admin.git --depth 1 > /dev/null 2>&1"
             )
@@ -116,7 +116,7 @@ def main(argv=sys.argv[1:]):
                     "RUN install bulkhours_admin: installation failed 🚫. Check that your atoken is still valid (contact: bulkhours@guydegnol.net)"
                 )
 
-        if args.tokens["mtoken"] != DEFAULT_TOKEN:
+        if "mtoken" in args.tokens and args.tokens["mtoken"] != DEFAULT_TOKEN:
             os.system(
                 f"cd {bulk_dir} && rm -rf bulkhours_premium 2> /dev/null && git clone https://{args.tokens['mtoken']}@github.com/guydegnol/bulkhours_premium.git --depth 1 > /dev/null 2>&1"
             )
